backend-1/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
PyJWT
python-multipart
email-validator
python-dotenv
requests
